test extensions in their own packages
